fastapi-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pyyaml
python-dotenv>=1.0.0
openai>=1.12.0
requests>=2.31.0
python-docx>=1.1.0
pypdf>=4.0.0
python-dateutil>=2.8.2
pandas>=2.1.0
openpyxl>=3.1.0
networkx>=3.0
plotly>=5.0.0
gnews>=0.3.0
googlenewsdecoder>=0.1.4
trafilatura>=1.6.0
newspaper3k>=0.2.8
rapidfuzz>=3.5.0
pytz
urllib3<2
